requirements.txt
flask==2.0.1
nltk==3.8.1
scikit-learn==1.3.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
torch==2.0.1
transformers==4.30.2
tqdm==4.65.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
vaderSentiment==3.3.2
Flask-MySQL==1.5.2
mysql-connector-python==8.0.26
python-Levenshtein==0.21.1
spacy==3.5.3
textblob==0.17.1
huggingface-hub==0.16.4
Werkzeug==2.0.1
requests==2.26.0
python-docx==0.8.11
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.0.0
opencv-python==4.6.0.66
PyMuPDF==1.20.0
paddlepaddle==2.5.2
paddleocr==2.7.0.3
mysqlclient==2.0.3 